Make InteractiveBase abstract
Not something that's a massive problem but earlier in the Discord.Net channel someone was trying to create a new instance of `InteractiveBase` and access its methods. This change just makes it more consistent with `ModuleBase` and `ModuleBase<T>` but also explicitly shows that the class is intended to be inherited.