Added more control over how a paged message gets handled
Inside the PaginatedMessage you can now set if the message should be forced to be send as an Embed or just keep it as is. You can also now disable any edits that the library makes to the embed, e.g. disable the page count at the bottom of the embed